refactor(LineSectionAddAcceptanceTest.java): 구간 등록 인수 테스트 클래스 이름 변경 및 패키지 이동
구간 등록 인수 테스트 클래스 이름 변경 및 패키지 이동
- LineSectionAddAcceptanceTest : SectionAcceptanceTest -> LineSectionAddAcceptanceTest 이름 변경
- LineSectionAddAcceptanceMethods : SectionAcceptanceMethods -> LineSectionAddAcceptanceMethods 이름 변경
