+++ src/test/java/nextstep/subway/line/LineSectionAddAcceptanceMethods.java
@@ -1,4 +1,4 @@
-package nextstep.subway.section;
+package nextstep.subway.line;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -10,11 +10,11 @@ import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
 import nextstep.subway.section.dto.SectionRequest;
 
-public class SectionAcceptanceMethods extends AcceptanceTest {
+public class LineSectionAddAcceptanceMethods extends AcceptanceTest {
     private static final String LINE_URL_PATH = "/lines/";
     private static final String SECTION_URL_PATH = "/sections";
 
-    public SectionAcceptanceMethods() {}
+    public LineSectionAddAcceptanceMethods() {}
 
     public static ExtractableResponse<Response> 지하철_노선에_지하철역_등록_요청(Long lineId, SectionRequest sectionRequest) {
         return post(LINE_URL_PATH + lineId + SECTION_URL_PATH, sectionRequest);
